refactor: centralize shared common utilities
