+++ packages/backend/src/models/db.js
@@ -3,7 +3,7 @@ const pathUtil = require("../utils/path-util");
 const { getDirName } = pathUtil;
 const path = require('path');
 const serverUtil = require("../utils/server-util");
-const util = require('../common/util');
+const util = require('../../../common/src/util');
 const _util = require('util');
 const { isImage, isCompress, isMusic, isVideo, getCurrentTime } = util;
 const pfs = require('promise-fs');
