+++ packages/backend/src/routes/all-info.js
@@ -3,7 +3,7 @@ const express = require('express');
 const router = express.Router();
 const serverUtil = require("../utils/server-util");
 const db = require("../models/db");
-const util = require('../common/util');
+const util = require('../../../common/src/util');
 // const { isDisplayableInExplorer } = util;
 const nameParser = require('../name-parser');
 const _ = require("underscore");
