+++ packages/backend/src/routes/file-download.js
@@ -4,7 +4,7 @@ const pfs = require('promise-fs');
 const path = require('path');
 const express = require('express');
 const router = express.Router();
-const util = require('../common/util');
+const util = require('../../../common/src/util');
 const appState = require('../state/appState');
 const { isImage, isGif } = util;
 const serverUtil = require("../utils/server-util");
