+++ packages/backend/src/services/seven-zip.js
@@ -4,7 +4,7 @@ const pfs = require('promise-fs');
 const _ = require('underscore');
 const zipInfoDb = require("../models/zip-info-db");
 const logger = require("../config/logger");
-const util = require('../common/util');
+const util = require('../../../common/src/util');
 const pathUtil = require("../utils/path-util");
 // const { isImage, isCompress, isMusic, isVideo, getCurrentTime, isDisplayableInExplorer, isDisplayableInOnebook } = util;
 const { isExist } = pathUtil;
