+++ packages/backend/src/services/thumbnail-query.js
@@ -2,7 +2,7 @@ const path = require('path');
 const _ = require('underscore');
 
 const serverUtil = require("../utils/server-util");
-const util = require('../common/util');
+const util = require('../../../common/src/util');
 const { getCurrentTime } = util;
 
 const pathUtil = require("../utils/path-util");
