+++ packages/backend/src/utils/image-compress.js
@@ -1,6 +1,6 @@
 // const fs = require('fs');
 // const path = require('path');
-const util = require('../common/util');
+const util = require('../../../common/src/util');
 const execa = require('./own-execa');
 const pathUtil = require("./path-util");
 const logger = require("../config/logger");
