--- packages/frontend/src/common/constant.js
@@ -1,5 +1,0 @@
-module.exports.MODE_TAG = "mode tag";
-module.exports.MODE_AUTHOR = "mode author";
-module.exports.MODE_SEARCH = "mode search";
-module.exports.MODE_EXPLORER = "mode explorer";
-
